+++ woodwork/statistics_utils/frequency_inference/__init__.py
@@ -1,0 +1,1 @@
+from ._infer_frequency import infer_frequency
