+++ woodwork/statistics_utils/frequency_inference/_generate_estimated_timeseries.py
