--- woodwork/utils2/frequency_inference/__init__.py
@@ -1,1 +1,0 @@
-from .infer_frequency import infer_frequency
